@@ -6,12 +6,12 @@ android {
     buildFeatures {
         viewBinding true
     }
-    compileSdk 30
+    compileSdk 31
 
     defaultConfig {
         applicationId "com.astar.expirationdatemanagement"
-        minSdk 21
-        targetSdk 30
+        minSdk 26
+        targetSdk 31
         versionCode 1
         versionName "1.0"
 
@@ -31,17 +31,18 @@ android {
 }
 
 dependencies {
+    //implementation 'androidx.work:work-runtime:2.7.1'
     def roomVersion = "2.3.0"
     implementation("androidx.room:room-runtime:$roomVersion")
     annotationProcessor("androidx.room:room-compiler:$roomVersion")
 
     implementation 'com.journeyapps:zxing-android-embedded:4.3.0'
 
-    implementation 'androidx.appcompat:appcompat:1.3.1'
+    implementation 'androidx.appcompat:appcompat:1.4.0'
     implementation 'com.google.android.material:material:1.4.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.2'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    testImplementation 'junit:junit:4.+'
+    testImplementation 'junit:junit:'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 }
